+++ api/testsrc/no/mnemonic/act/platform/api/request/v1/RetractFactRequestTest.java
@@ -16,21 +16,23 @@ public class RetractFactRequestTest extends AbstractRequestTest {
   public void testDecodeRequest() throws Exception {
     UUID fact = UUID.randomUUID();
     UUID organization = UUID.randomUUID();
-    UUID source = UUID.randomUUID();
+    UUID origin = UUID.randomUUID();
     UUID acl = UUID.randomUUID();
     String json = String.format("{" +
             "fact : '%s'," +
             "organization : '%s'," +
-            "source : '%s'," +
+            "origin : '%s'," +
+            "confidence : 0.1," +
             "accessMode : 'Explicit'," +
             "comment : 'comment'," +
             "acl : ['%s']" +
-            "}", fact, organization, source, acl);
+            "}", fact, organization, origin, acl);
 
     RetractFactRequest request = getMapper().readValue(json, RetractFactRequest.class);
     assertEquals(fact, request.getFact());
     assertEquals(organization, request.getOrganization());
-    assertEquals(source, request.getSource());
+    assertEquals(origin, request.getOrigin());
+    assertEquals(0.1f, request.getConfidence(), 0.0);
     assertEquals(AccessMode.Explicit, request.getAccessMode());
     assertEquals("comment", request.getComment());
     assertEquals(ListUtils.list(acl), request.getAcl());
@@ -41,6 +43,26 @@ public class RetractFactRequestTest extends AbstractRequestTest {
     Set<ConstraintViolation<RetractFactRequest>> violations = getValidator().validate(new RetractFactRequest());
     assertEquals(1, violations.size());
     assertPropertyInvalid(violations, "fact");
+  }
+
+  @Test
+  public void testRequestValidationFailsOnMin() {
+    Set<ConstraintViolation<RetractFactRequest>> violations = getValidator().validate(new RetractFactRequest()
+            .setFact(UUID.randomUUID())
+            .setConfidence(-0.1f)
+    );
+    assertEquals(1, violations.size());
+    assertPropertyInvalid(violations, "confidence");
+  }
+
+  @Test
+  public void testRequestValidationFailsOnMax() {
+    Set<ConstraintViolation<RetractFactRequest>> violations = getValidator().validate(new RetractFactRequest()
+            .setFact(UUID.randomUUID())
+            .setConfidence(1.1f)
+    );
+    assertEquals(1, violations.size());
+    assertPropertyInvalid(violations, "confidence");
   }
 
   @Test